# web/src/api/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "API for the Rack Gateway administration and proxy services.",
        "title": "Rack Gateway API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api-tokens": {
            "get": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "Returns all API tokens configured in the system.",
                "tags": [
                    "API Tokens"
                ],
                "summary": "List API tokens",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/db.APIToken"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "SessionCookie": []
                    },
                    {
                        "CSRFToken": []
                    }
                ],
                "description": "Generates a new API token for automation or CI/CD use.",
                "tags": [
                    "API Tokens"
                ],
                "summary": "Create an API token",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/handlers.CreateAPITokenRequest"
                            }
                        }
                    },
                    "description": "Token payload",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.CreateAPITokenResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api-tokens/permissions": {
            "get": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "Returns the permission catalog and metadata used to build API token forms.",
                "tags": [
                    "API Tokens"
                ],
                "summary": "Get token permission metadata",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.TokenPermissionMetadata"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api-tokens/{tokenID}": {
            "get": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "Returns metadata for a specific API token.",
                "tags": [
                    "API Tokens"
                ],
                "summary": "Get API token",
                "parameters": [
                    {
                        "description": "Token ID",
                        "name": "tokenID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/db.APIToken"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "SessionCookie": []
                    },
                    {
                        "CSRFToken": []
                    }
                ],
                "description": "Updates token metadata such as name and permissions.",
                "tags": [
                    "API Tokens"
                ],
                "summary": "Update an API token",
                "parameters": [
                    {
                        "description": "Token ID",
                        "name": "tokenID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/handlers.UpdateAPITokenRequest"
                            }
                        }
                    },
                    "description": "Token update",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/db.APIToken"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "SessionCookie": []
                    },
                    {
                        "CSRFToken": []
                    }
                ],
                "description": "Permanently removes an API token.",
                "tags": [
                    "API Tokens"
                ],
                "summary": "Delete an API token",
                "parameters": [
                    {
                        "description": "Token ID",
                        "name": "tokenID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "components": {
...
        "securitySchemes": {
            "CSRFToken": {
                "description": "HMAC-derived CSRF token tied to the active session.",
                "type": "apiKey",
                "name": "X-CSRF-Token",
                "in": "header"
            },
            "SessionCookie": {
                "description": "HttpOnly session cookie issued after OAuth login.",
                "type": "apiKey",
                "name": "Cookie",
                "in": "header"
            }
        },
        "schemas": {
...
            "db.APIToken": {
                "type": "object",
                "properties": {
                    "created_at": {
                        "type": "string"
                    },
                    "created_by_email": {
                        "type": "string"
                    },
                    "created_by_name": {
                        "type": "string"
                    },
                    "created_by_user_id": {
                        "type": "integer"
                    },
                    "expires_at": {
                        "type": "string"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "last_used_at": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "permissions": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "public_id": {
                        "type": "string"
                    },
                    "user_id": {
                        "type": "integer"
                    }
                }
            },
...
            "handlers.CreateAPITokenRequest": {
                "type": "object",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "permissions": {
                        "description": "Explicit permissions (overrides role)",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "role": {
                        "description": "Role shortcut (viewer, ops, deployer, cicd, admin)",
                        "type": "string"
                    },
                    "user_email": {
                        "type": "string"
                    }
                }
            },
            "handlers.CreateAPITokenResponse": {
                "type": "object",
                "required": [
                    "api_token",
                    "id",
                    "name",
                    "permissions",
                    "token"
                ],
                "properties": {
                    "api_token": {
                        "$ref": "#/components/schemas/db.APIToken"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "permissions": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "token": {
                        "type": "string"
                    }
                }
            },
...
            "handlers.ErrorResponse": {
                "type": "object",
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "type": "string"
                    }
                }
            },
...
            "handlers.RoleDescriptor": {
                "type": "object",
                "required": [
                    "description",
                    "label",
                    "name",
                    "permissions"
                ],
                "properties": {
                    "description": {
                        "type": "string"
                    },
                    "label": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "permissions": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
...
            "handlers.TokenPermissionMetadata": {
                "type": "object",
                "required": [
                    "default_permissions",
                    "permissions",
                    "roles",
                    "user_permissions",
                    "user_roles"
                ],
                "properties": {
                    "default_permissions": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "permissions": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/handlers.RoleDescriptor"
                        }
                    },
                    "user_permissions": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "user_roles": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
...
            "handlers.UpdateAPITokenRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "permissions": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
...
        }
    }
}